requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
pydantic
pydantic-settings
PyJWT
passlib
bcrypt==3.2.2
python-multipart
apscheduler
pytest
httpx